tests: refactor tests to allow registration of bp with custom services
